requirements.txt
fastapi
uvicorn
langchain-google-genai
pillow
python-multipart
python-dotenv
